pyproject.toml
# SPDX-License-Identifier: Apache-2.0
# This file was modified with the assistance of an AI (Large Language Model). Review required for correctness, security, and licensing.

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rackwright"
version = "0.1.0"
description = "Rackwright implementation"
requires-python = ">=3.11"
dependencies = [
  "Flask>=3.0,<4.0",
  "SQLAlchemy>=2.0,<3.0",
  "alembic>=1.13,<2.0",
  "python-docx>=1.1,<2.0",
  "openpyxl>=3.1,<4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "ruff>=0.9,<1.0",
  "black>=24.0,<26.0",
  "mypy>=1.11,<2.0",
  "pre-commit>=3.8,<5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
pythonpath = ["."]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
exclude = ["alembic/versions/"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["rackwright*"]

[tool.setuptools.package-data]
rackwright = ["templates/*.html"]
